[CHORE] week5 + 회원 탈퇴 머지
