+++ src/main/java/channeling/be/global/auth/handler/Oauth2LoginSuccessHandler.java
@@ -1,5 +1,6 @@
 package channeling.be.global.auth.handler;
 
+import channeling.be.domain.member.domain.Member;
 import channeling.be.domain.member.domain.repository.MemberRepository;
 import channeling.be.global.auth.application.MemberOauth2UserService;
 import channeling.be.global.auth.application.MemberOauth2UserService.LoginResult;
@@ -20,6 +21,7 @@ import org.springframework.web.util.UriComponentsBuilder;
 
 import java.io.IOException;
 import java.util.Map;
+import java.util.Optional;
 
 
 @Slf4j
@@ -63,8 +65,6 @@ public class Oauth2LoginSuccessHandler implements AuthenticationSuccessHandler {
         System.out.println("Refresh Token: " + googleRefreshToken);
 
 
-
-
         log.info("컨텍스트에서 구글 엑세스 토큰 추출 = {}", googleAccessToken);
         /* -------------------------------------------------
          * OAuth2User 꺼내기
@@ -77,23 +77,25 @@ public class Oauth2LoginSuccessHandler implements AuthenticationSuccessHandler {
         String googleId = attrs.get("sub").toString();
 
         // 3. 회원 탈퇴 여부 확인
-        memberRepository.findByGoogleId(googleId).ifPresent(member -> {
-            if (member.getIsDeleted()) {
+        Optional<Member> memberOp = memberRepository.findByGoogleId(googleId);
 
-                // 프론트 응답 생성
-                String targetUrl = UriComponentsBuilder.fromUriString("http://localhost:5173/auth/callback")
-                        .queryParam("token", jwtUtil.createTempToken(member)) // 재활성화를 위한 임시 토큰 발급
-                        .queryParam("message", "Deleted member")
-                        .build()
-                        .toUriString();
+        if (memberOp.isPresent() && memberOp.get().getIsDeleted()) {
 
-                try {
-                    response.sendRedirect(targetUrl);
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
-                }
+            // 프론트 응답 생성
+            String targetUrl = UriComponentsBuilder.fromUriString("http://localhost:5173/auth/callback")
+                    .queryParam("token", jwtUtil.createTempToken(memberOp.get())) // 재활성화를 위한 임시 토큰 발급
+                    .queryParam("message", "Deleted member")
+                    .build()
+                    .toUriString();
+
+            try {
+                response.sendRedirect(targetUrl);
+                return;
+            } catch (IOException e) {
+                throw new RuntimeException(e);
             }
-        });
+        }
+
 
         log.info("컨텍스트에서 토큰 내의 회원 정보 추출 = {}", attrs);
 
@@ -117,8 +119,8 @@ public class Oauth2LoginSuccessHandler implements AuthenticationSuccessHandler {
         String accessToken = jwtUtil.createAccessToken(result.member());
         String refreshToken = jwtUtil.createRefreshToken(result.member());
 
-         log.info("Access Token: {}", accessToken);
-         log.info("Refresh Token: {}", refreshToken);
+        log.info("Access Token: {}", accessToken);
+        log.info("Refresh Token: {}", refreshToken);
         // 프론트 응답 생성
         String targetUrl = UriComponentsBuilder.fromUriString(frontUrl + "/auth/callback") // TODO
                 .queryParam("token", accessToken)
